fix(tests): split process-next claim races onto workerless factory (M3)
