+++ backend/tests/Taskdeck.Api.Tests/Concurrency/QueueClaimRaceTests.cs
@@ -24,6 +24,12 @@ namespace Taskdeck.Api.Tests.Concurrency;
 /// guards would prevent true concurrent claim races at the row level.
 ///
 /// See GitHub issue #705 (TST-55).
+///
+/// Worker split (issue #1335): this class keeps the base worker-RUNNING factory because its
+/// capture-triage tests DEPEND on the live triage drain (they poll until the worker turns the
+/// triaged capture into a proposal). The process-next claim races — which the live
+/// <c>LlmQueueToProposalWorker</c> can pre-empt by draining the seeded Pending row first — moved
+/// to <see cref="ProcessNextClaimRaceTests"/> on the workerless factory.
 /// </summary>
 public class QueueClaimRaceTests : IClassFixture<TestWebApplicationFactory>
 {
@@ -32,69 +38,6 @@ public class QueueClaimRaceTests : IClassFixture<TestWebApplicationFactory>
     public QueueClaimRaceTests(TestWebApplicationFactory factory)
     {
         _factory = factory;
-    }
-
-    /// <summary>
-    /// Scenario 1: 10 parallel workers all try to process the next pending LLM queue item.
-    /// Under SQLite's file-level write serialization, multiple workers may read the same
-    /// pending item before any status update commits, causing more than one to succeed.
-    /// This is a known SQLite limitation -- with PostgreSQL row-level locking, at most
-    /// one worker would claim each item.
-    ///
-    /// The test validates:
-    /// - No 500 errors under concurrent access
-    /// - At least one worker succeeds
-    /// - No deadlocks or hangs (test completes within timeout)
-    /// </summary>
-    [Fact]
-    public async Task ProcessNext_TenParallelWorkers_NoErrorsUnderConcurrentAccess()
-    {
-        const int workerCount = 10;
-        using var setupClient = _factory.CreateClient();
-        await ApiTestHarness.AuthenticateAsync(setupClient, "queue-claim-10");
-
-        // Seed a single LLM queue item
-        var queueResp = await setupClient.PostAsJsonAsync(
-            "/api/llm-queue",
-            new CreateLlmRequestDto("summarize", "payload for claim race"));
-        queueResp.StatusCode.Should().Be(HttpStatusCode.OK);
-
-        // Fire 10 parallel process-next requests
-        using var barrier = new SemaphoreSlim(0, workerCount);
-        var results = new ConcurrentBag<HttpStatusCode>();
-
-        var tasks = Enumerable.Range(0, workerCount).Select(async _ =>
-        {
-            using var workerClient = _factory.CreateClient();
-            workerClient.DefaultRequestHeaders.Authorization =
-                setupClient.DefaultRequestHeaders.Authorization;
-            await barrier.WaitAsync();
-            var resp = await workerClient.PostAsync("/api/llm-queue/process-next", null);
-            results.Add(resp.StatusCode);
-        }).ToArray();
-
-        barrier.Release(workerCount);
-        await Task.WhenAll(tasks);
-
-        var codes = results.ToList();
-        var successCount = codes.Count(s => s == HttpStatusCode.OK);
-
-        // At least one worker should succeed
-        successCount.Should().BeGreaterOrEqualTo(1,
-            "at least one worker should process the pending item");
-
-        // NOTE: Under SQLite, multiple workers may succeed because reads are not
-        // serialized against writes at the row level. With PostgreSQL, we would
-        // expect successCount <= 1 due to SELECT ... FOR UPDATE or advisory locks.
-        // The important invariant is no 500 errors and no deadlocks.
-
-        // Remaining workers should get 404 (no pending item) or OK.
-        // Under SQLite's file-level write lock, transient 500s from DB contention
-        // are a known test-environment limitation and are tolerated.
-        codes.Should().OnlyContain(
-            s => s == HttpStatusCode.OK || s == HttpStatusCode.NotFound
-                 || s == HttpStatusCode.BadRequest || s == HttpStatusCode.InternalServerError,
-            "workers should only get OK, 404, 400, or transient 500 from SQLite contention");
     }
 
     /// <summary>
@@ -235,57 +178,4 @@ public class QueueClaimRaceTests : IClassFixture<TestWebApplicationFactory>
         }
     }
 
-    /// <summary>
-    /// Scenario: Two workers both call process-next simultaneously for different
-    /// pending items. Each should claim a different item (no double processing).
-    /// </summary>
-    [Fact]
-    public async Task ProcessNext_TwoWorkersTwoItems_EachClaimsDifferentItem()
-    {
-        using var setupClient = _factory.CreateClient();
-        await ApiTestHarness.AuthenticateAsync(setupClient, "queue-two-workers");
-
-        // Seed two LLM queue items
-        var q1 = await setupClient.PostAsJsonAsync(
-            "/api/llm-queue",
-            new CreateLlmRequestDto("summarize", "payload-A"));
-        q1.StatusCode.Should().Be(HttpStatusCode.OK);
-
-        var q2 = await setupClient.PostAsJsonAsync(
-            "/api/llm-queue",
-            new CreateLlmRequestDto("summarize", "payload-B"));
-        q2.StatusCode.Should().Be(HttpStatusCode.OK);
-
-        // Two workers process-next simultaneously
-        using var barrier = new SemaphoreSlim(0, 2);
-        var responseData = new ConcurrentBag<(HttpStatusCode Status, string? Body)>();
-
-        var workerTasks = Enumerable.Range(0, 2).Select(async _ =>
-        {
-            using var workerClient = _factory.CreateClient();
-            workerClient.DefaultRequestHeaders.Authorization =
-                setupClient.DefaultRequestHeaders.Authorization;
-            await barrier.WaitAsync();
-            var resp = await workerClient.PostAsync("/api/llm-queue/process-next", null);
-            var body = await resp.Content.ReadAsStringAsync();
-            responseData.Add((resp.StatusCode, body));
-        }).ToArray();
-
-        barrier.Release(2);
-        await Task.WhenAll(workerTasks);
-
-        var responses = responseData.ToList();
-
-        // At least one worker should succeed (with 2 items, ideally both succeed)
-        var successResponses = responses.Where(r => r.Status == HttpStatusCode.OK).ToList();
-        successResponses.Should().NotBeEmpty(
-            "at least one worker should successfully claim an item");
-
-        // All responses should be well-formed. Under SQLite's file-level write lock,
-        // transient 500s from DB contention are a known test-environment limitation.
-        responses.Should().OnlyContain(
-            r => r.Status == HttpStatusCode.OK || r.Status == HttpStatusCode.NotFound
-                 || r.Status == HttpStatusCode.BadRequest || r.Status == HttpStatusCode.InternalServerError,
-            "workers should only get OK, 404, 400, or transient 500 from SQLite contention");
-    }
 }
